+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 **2020-07: 3.9.7**
 
+- resolve dedendabot alert about 'qs' by replacing gulp-webserver with gulp-connect + open
+  
 - resolve #138 based on @lewismc's https://github.com/mmisw/orr-portal/pull/160 PR.
   Many thanks Lewis! I went ahead and manually applied the key changes from your PR
   as there have been quite a few changes in related files (and didn't want to deal
@@ -10,9 +12,6 @@
   any lodash.template stuff.
   Good testing with `gulp try_dist --localConfig --base=/orrportal/`.
 
-- in `gulpfile.js` also fixed a bad location of `local.config.js`, which, per some
-  other recent adjustments, is now to be located under `src/app/config/`.
-  
 **2020-07: 3.9.6**
 
 - adjust version after reverting unneeded change