+++ ChangeLog.md
@@ -1,5 +1,9 @@
 ## change log ##
 
+**2020-07: 3.9.8**
+
+- fix #174 "No URL-encoding when dispatching external viewer for fully-hosted ontology"
+
 **2020-07: 3.9.7**
 
 - adjust marks for status and visibility, in particular to show the visibility